+++ src/main/java/pl/alvion/rpem/rpessentials/RPEssentials.java
@@ -7,7 +7,7 @@ import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.plugin.java.JavaPlugin;
 import pl.alvion.rpem.rpessentials.commands.GeneralDebugCommand;
 import pl.alvion.rpem.rpessentials.generalutils.gui.GUIListener;
-import pl.alvion.rpem.rpessentials.rpplayer.RPPlayerListener;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.RPPlayerListener;
 
 import java.io.File;
 import java.io.IOException;
