+++ src/main/java/pl/alvion/rpem/rpessentials/rpplayer_dr/stats/aviliable/Agility.java
@@ -1,24 +1,23 @@
-package pl.alvion.rpem.rpessentials.rpplayer.stats.aviliable;
+package pl.alvion.rpem.rpessentials.rpplayer_dr.stats.aviliable;
 
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
-import pl.alvion.rpem.rpessentials.rpplayer.stats.Stat;
-import pl.alvion.rpem.rpessentials.rpplayer.stats.Stats;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.stats.Stat;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.stats.Stats;
 
-public class MaxHP extends Stat {
-
+public class Agility extends Stat {
 
 
     @Override
     public String name() {
-        return ChatColor.LIGHT_PURPLE + "Zdrowie";
+        return ChatColor.GRAY + "Zrecznosc";
     }
 
     @Override
     public ItemStack guiItem() {
-        return new ItemStack(Material.PINK_DYE, 1);
+        return new ItemStack(Material.FEATHER, 1);
     }
 
     @Override
@@ -28,6 +27,8 @@ public class MaxHP extends Stat {
 
     @Override
     public Stats statName() {
-        return Stats.MaxHP;
+        return Stats.Agility;
     }
+
+
 }
