+++ src/main/java/pl/alvion/rpem/rpessentials/rpplayer_dr/stats/aviliable/Endurance.java
@@ -1,24 +1,23 @@
-package pl.alvion.rpem.rpessentials.rpplayer.stats.aviliable;
+package pl.alvion.rpem.rpessentials.rpplayer_dr.stats.aviliable;
 
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
-import pl.alvion.rpem.rpessentials.rpplayer.stats.Stat;
-import pl.alvion.rpem.rpessentials.rpplayer.stats.Stats;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.stats.Stat;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.stats.Stats;
 
-public class Magic extends Stat {
+public class Endurance extends Stat {
 
     @Override
     public String name() {
-        return ChatColor.DARK_PURPLE + "Potencjal Magiczny";
+        return ChatColor.DARK_GRAY + "Wytrzymalosc";
     }
 
     @Override
     public ItemStack guiItem() {
-        return new ItemStack(Material.NETHER_STAR, 1);
+        return new ItemStack(Material.IRON_CHESTPLATE);
     }
-
 
     @Override
     public boolean canUpgrade(Player player) {
@@ -27,6 +26,6 @@ public class Magic extends Stat {
 
     @Override
     public Stats statName() {
-        return Stats.Magic;
+        return Stats.Endurance;
     }
 }
