+++ src/main/java/pl/alvion/rpem/rpessentials/rpplayer_dr/stats/aviliable/Intelligence.java
@@ -1,22 +1,23 @@
-package pl.alvion.rpem.rpessentials.rpplayer.stats.aviliable;
+package pl.alvion.rpem.rpessentials.rpplayer_dr.stats.aviliable;
 
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
-import pl.alvion.rpem.rpessentials.rpplayer.stats.Stat;
-import pl.alvion.rpem.rpessentials.rpplayer.stats.Stats;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.stats.Stat;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.stats.Stats;
 
-public class Endurance extends Stat {
+public class Intelligence extends Stat {
+
 
     @Override
     public String name() {
-        return ChatColor.DARK_GRAY + "Wytrzymalosc";
+        return ChatColor.BLUE + "Inteligencja";
     }
 
     @Override
     public ItemStack guiItem() {
-        return new ItemStack(Material.IRON_CHESTPLATE);
+        return new ItemStack(Material.BOOK, 1);
     }
 
     @Override
@@ -26,6 +27,6 @@ public class Endurance extends Stat {
 
     @Override
     public Stats statName() {
-        return Stats.Endurance;
+        return Stats.Intelligence;
     }
 }
