+++ src/main/java/pl/alvion/rpem/rpessentials/rpplayer_dr/stats/aviliable/MaxHP.java
@@ -1,23 +1,24 @@
-package pl.alvion.rpem.rpessentials.rpplayer.stats.aviliable;
+package pl.alvion.rpem.rpessentials.rpplayer_dr.stats.aviliable;
 
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
-import pl.alvion.rpem.rpessentials.rpplayer.stats.Stat;
-import pl.alvion.rpem.rpessentials.rpplayer.stats.Stats;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.stats.Stat;
+import pl.alvion.rpem.rpessentials.rpplayer_dr.stats.Stats;
 
-public class Intelligence extends Stat {
+public class MaxHP extends Stat {
+
 
 
     @Override
     public String name() {
-        return ChatColor.BLUE + "Inteligencja";
+        return ChatColor.LIGHT_PURPLE + "Zdrowie";
     }
 
     @Override
     public ItemStack guiItem() {
-        return new ItemStack(Material.BOOK, 1);
+        return new ItemStack(Material.PINK_DYE, 1);
     }
 
     @Override
@@ -27,6 +28,6 @@ public class Intelligence extends Stat {
 
     @Override
     public Stats statName() {
-        return Stats.Intelligence;
+        return Stats.MaxHP;
     }
 }
